+++ app/src/main/java/com/codesoom/assignment/product/dto/ProductData.java
@@ -7,7 +7,6 @@ import lombok.NoArgsConstructor;
 
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
-import java.math.BigDecimal;
 
 @Getter
 @NoArgsConstructor
@@ -24,13 +23,13 @@ public class ProductData {
 
     @NotNull
     @Mapping("price")
-    private BigDecimal price;
+    private long price;
 
     @Mapping("imageUrl")
     private String imageUrl;
 
     @Builder
-    public ProductData(Long id, String name, String maker, BigDecimal price, String imageUrl) {
+    public ProductData(Long id, String name, String maker, long price, String imageUrl) {
         this.id = id;
         this.name = name;
         this.maker = maker;
